+++ src/client/component/weapon.py
@@ -1,9 +1,10 @@
 from dataclasses import dataclass
 
+
 @dataclass
 class Weapon:
-    cooldown_max : float
-    bullet_speed : float
-    damage : float
-    
-    cooldown_current : float = 0.0
+    cooldown_max: float
+    bullet_speed: float
+    damage: int
+
+    cooldown_current: float = 0.0
